Feature/finished/IIA-2874: [Pattern Stamp Editing] New version of Stamp Edit mode for Pattern Window
