+++ kview/src/main/java/dev/ikm/komet/kview/mvvm/view/common/StampFormController.java
@@ -1,9 +1,9 @@
 package dev.ikm.komet.kview.mvvm.view.common;
 
-import static dev.ikm.komet.kview.mvvm.viewmodel.StampFormViewModelBase.StampProperties.*;
+import static dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampFormViewModelBase.StampProperties.*;
 import dev.ikm.komet.framework.view.*;
 import dev.ikm.komet.kview.common.ViewCalculatorUtils;
-import dev.ikm.komet.kview.mvvm.viewmodel.*;
+import dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampFormViewModelBase;
 import dev.ikm.tinkar.terms.*;
 import javafx.beans.binding.StringBinding;
 import javafx.beans.property.*;
