+++ kview/src/main/java/dev/ikm/komet/kview/mvvm/view/pattern/PropertiesController.java
@@ -17,6 +17,9 @@ package dev.ikm.komet.kview.mvvm.view.pattern;
 
 import dev.ikm.komet.kview.events.StampEvent;
 import dev.ikm.komet.kview.mvvm.view.common.StampFormController;
+import dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampAddConfirmFormViewModel;
+import dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampCreateFormViewModel;
+import dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampFormViewModelBase;
 import dev.ikm.tinkar.events.EvtBusFactory;
 import dev.ikm.tinkar.events.EvtType;
 import dev.ikm.tinkar.events.Subscriber;
@@ -65,7 +68,7 @@ import static dev.ikm.komet.kview.mvvm.viewmodel.DescrNameViewModel.MODE;
 import static dev.ikm.komet.kview.mvvm.viewmodel.DescrNameViewModel.VIEW_PROPERTIES;
 import static dev.ikm.komet.kview.mvvm.viewmodel.PatternFieldsViewModel.*;
 import static dev.ikm.komet.kview.mvvm.viewmodel.PatternViewModel.*;
-import static dev.ikm.komet.kview.mvvm.viewmodel.StampFormViewModelBase.StampType.PATTERN;
+import static dev.ikm.komet.kview.mvvm.viewmodel.stamp.StampFormViewModelBase.StampType.PATTERN;
 import static dev.ikm.komet.kview.state.PatternDetailsState.NEW_PATTERN_INITIAL;
 import static dev.ikm.tinkar.terms.TinkarTerm.FULLY_QUALIFIED_NAME_DESCRIPTION_TYPE;
 import static dev.ikm.tinkar.terms.TinkarTerm.REGULAR_NAME_DESCRIPTION_TYPE;
@@ -161,13 +164,13 @@ public class PropertiesController {
 
     private Subscriber<StampEvent> createStampSubscriber;
 
-    private StampAddFormViewModel stampAddFormViewModel;
+    private StampAddConfirmFormViewModel stampAddConfirmFormViewModel;
     private StampCreateFormViewModel stampCreateFormViewModel;
 
     private EntityFacade patternFacade;
 
     public PropertiesController() {
-        this.stampAddFormViewModel = new StampAddFormViewModel(PATTERN);
+        this.stampAddConfirmFormViewModel = new StampAddConfirmFormViewModel(PATTERN);
         this.stampCreateFormViewModel = new StampCreateFormViewModel(PATTERN);
     }
 
@@ -367,7 +370,7 @@ public class PropertiesController {
         // -- add stamp
         addStampSubscriber = evt -> {
             if (evt.getEventType() == StampEvent.ADD_STAMP) {
-                stampJFXNode.controller().init(stampAddFormViewModel);
+                stampJFXNode.controller().init(stampAddConfirmFormViewModel);
                 contentBorderPane.setCenter(stampJFXNode.node());
             }
         };
@@ -388,8 +391,8 @@ public class PropertiesController {
     public void updateModel(EntityFacade newPattern) {
         this.patternFacade = newPattern;
 
-        if (newPattern != null && stampAddFormViewModel != null) {
-            setStampFormViewModel(stampAddFormViewModel);
+        if (newPattern != null && stampAddConfirmFormViewModel != null) {
+            setStampFormViewModel(stampAddConfirmFormViewModel);
         } else if (newPattern == null && stampCreateFormViewModel != null) {
             setStampFormViewModel(stampCreateFormViewModel);
         }
