+++ kview/src/main/java/module-info.java
@@ -156,6 +156,8 @@ module dev.ikm.komet.kview {
 
     exports dev.ikm.komet.kview.mvvm.view.loginauthor;
     opens dev.ikm.komet.kview.mvvm.view.loginauthor;
+    exports dev.ikm.komet.kview.mvvm.viewmodel.stamp;
+    opens dev.ikm.komet.kview.mvvm.viewmodel.stamp to dev.ikm.komet.application, javafx.fxml, org.carlfx.cognitive;
 
     provides dev.ikm.komet.framework.KometNodeFactory with dev.ikm.komet.kview.mvvm.view.concept.ConceptNodeFactory;
